backend/requirements.txt
# ─── Core Framework ───────────────────────────────────────────
fastapi==0.136.1
uvicorn[standard]==0.38.0
pydantic==2.13.3
pydantic-settings==2.12.0

# ─── AI / ML ─────────────────────────────────────────────────
langchain-text-splitters==0.3.8
sentence-transformers==4.1.0
huggingface-hub==0.34.4
chromadb==1.5.8

# ─── LLM Providers ───────────────────────────────────────────
openai==1.78.1
anthropic==0.49.0
google-genai==1.74.0

# ─── Document Parsing ────────────────────────────────────────
pymupdf==1.25.4
python-docx==1.1.2
python-multipart==0.0.20

# ─── Database ────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.41
asyncpg==0.31.0
alembic==1.15.2

# ─── Authentication ──────────────────────────────────────────
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
email-validator==2.3.0
authlib==1.5.2
httpx==0.28.1

# ─── Utilities ───────────────────────────────────────────────
python-dotenv==1.1.0
